Moved mock of JobManager to job manager's file, an re-enabled some tests
